Project_finl/requirements.txt
networkx
matplotlib
pandas
scikit-learn
reportlab
python-docx
